requirements.txt
flask==2.2.5
pandas==2.2.3
scikit-learn==1.4.2
pycaret==3.3.2
python-dotenv==1.0.1
gunicorn==21.2.0
flask-cors==5.0.0 
numpy==1.26.4
matplotlib==3.7.1
seaborn==0.12.2
nltk==3.9.1
wordcloud==1.9.4